# backend/openapi/grupos.yaml
openapi: 3.0.3
info:
  title: Mesh API — Grupos
  version: 1.0.0
  description: |
    Endpoints de sincronización de usuario y gestión de grupos (SCRUM-11).

    Autenticación en endpoints protegidos: header `Authorization: Bearer <Firebase ID token>`.
    El endpoint `POST /api/usuarios/sync` es público (MVP) hasta integrar Firebase Auth en backend.

servers:
  - url: http://localhost:3000
    description: Desarrollo local

paths:
  /api/usuarios/sync:
    post:
      summary: Sincronizar usuario Firebase con PostgreSQL
      description: |
        Upsert por email. Devuelve el UUID backend usado como `x-user-id`.
        **Deuda técnica:** sin verificación de token Firebase.
      operationId: syncUsuario
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncUsuarioRequest'
      responses:
        '200':
          description: Usuario sincronizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncUsuarioResponse'
        '400':
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos:
    post:
      summary: Crear grupo
      description: |
        Crea un grupo y asigna al creador como líder (RN-010).
        También crea fila en `grupo_miembro` con rol `lider`.
      operationId: crearGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Grupo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrupoResponse'
        '400':
          description: Validación fallida (ej. nombre vacío)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}:
    get:
      summary: Detalle de grupo
      description: Solo accesible para miembros del grupo.
      operationId: obtenerGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle del grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrupoDetalleResponse'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Usuario no es miembro del grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Eliminar grupo
      description: |
        Elimina el grupo por completo (SCRUM-16). Solo el líder actual.
        Borra viajes planificados asociados; bloquea si hay un viaje en curso.
      operationId: eliminarGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Grupo eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrupoMutationResponse'
        '403':
          description: Solo el líder puede eliminar el grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Hay un viaje en curso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/abandonar:
    post:
      summary: Abandonar grupo
      description: |
        Remueve al usuario autenticado del grupo (SCRUM-16).
        Si es participante, sale directamente.
        Si es líder único, elimina el grupo.
        Si es líder con otros miembros, requiere `nuevo_lider_id` para transferir liderazgo.
      operationId: abandonarGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbandonarGrupoRequest'
      responses:
        '200':
          description: Abandono o eliminación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrupoMutationResponse'
        '400':
          description: Líder debe transferir liderazgo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Usuario no es miembro del grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo o integrante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Hay un viaje en curso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/miembros:
    get:
      summary: Listar miembros del grupo
      description: |
        Devuelve todos los integrantes con nombre, email y rol (SCRUM-15).
        Solo accesible para miembros del grupo.
      operationId: listarMiembrosGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de miembros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrupoMiembroResponse'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Usuario no es miembro del grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/miembros/{usuarioId}/rol:
    patch:
      summary: Cambiar rol de un integrante
      description: |
        Solo el líder actual puede cambiar roles (RN-012, RN-030).
        Promover a `lider` transfiere el liderazgo único del grupo.
      operationId: cambiarRolMiembroGrupo
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: usuarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarRolMiembroRequest'
      responses:
        '200':
          description: Rol actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CambiarRolMiembroResponse'
        '400':
          description: Validación fallida o operación no permitida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Solo el líder puede cambiar roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo o integrante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/invitaciones/pendientes:
    get:
      summary: Listar invitaciones pendientes del usuario
      description: |
        Devuelve invitaciones a grupos con estado `pendiente` para el usuario autenticado (SCRUM-17).
      operationId: listarInvitacionesPendientes
      tags:
        - Grupos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de invitaciones pendientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitacionPendienteResponse'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/invitaciones/{invitacionId}/responder:
    post:
      summary: Aceptar o rechazar invitación a grupo
      description: |
        Responde una invitación pendiente (SCRUM-17).
        Si `accion` es `aceptar`, marca la invitación como `aceptada` y agrega al usuario como `participante` del grupo.
        Si es `rechazar`, marca la invitación como `rechazada` sin modificar membresía.
      operationId: responderInvitacion
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: invitacionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponderInvitacionRequest'
      responses:
        '200':
          description: Invitación respondida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponderInvitacionResponse'
        '400':
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invitación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invitación ya respondida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/amigos-para-invitar:
    get:
      summary: Listar amigos invitables al grupo
      description: |
        Devuelve amistades aceptadas del líder con flag `ya_es_miembro` para el grupo destino (SCRUM-18).
      operationId: listarAmigosParaInvitar
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de amigos con estado de membresía
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioParaInvitarResponse'
        '403':
          description: Solo el líder puede invitar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/buscar-usuarios:
    get:
      summary: Buscar usuarios globales para invitar al grupo
      description: |
        Búsqueda por nombre (mín. 2 caracteres) en toda la base de usuarios (SCRUM-18).
      operationId: buscarUsuariosParaInvitar
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 80
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioParaInvitarResponse'
        '400':
          description: Query inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Solo el líder puede invitar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/usuarios-para-invitar:
    get:
      summary: Listar personas invitables al grupo
      description: |
        Devuelve integrantes de los otros grupos del líder, excluyendo miembros actuales,
        invitaciones pendientes y al propio líder (SCRUM-13).
      operationId: listarUsuariosParaInvitar
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de personas invitables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioInvitableResponse'
        '403':
          description: Solo el líder puede invitar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/invitar-usuarios:
    post:
      summary: Invitar personas seleccionadas al grupo
      description: |
        Crea invitaciones pendientes para los usuarios elegidos (SCRUM-13/SCRUM-18).
        Acepta amigos, usuarios de búsqueda global o co-miembros de otros grupos.
        `grupo_origen_id` es null cuando la invitación no proviene de otro grupo.
      operationId: invitarUsuarios
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitarUsuariosRequest'
      responses:
        '201':
          description: Invitaciones registradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitarUsuariosResponse'
        '400':
          description: Personas no invitables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Solo el líder puede invitar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/grupos-para-invitar:
    get:
      summary: Listar grupos disponibles para invitación masiva
      description: |
        Devuelve los demás grupos del usuario (excluye el grupo destino) con cantidad de integrantes (SCRUM-13).
        Solo el líder del grupo destino puede consultar la lista (RN-030).
      operationId: listarGruposParaInvitar
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de grupos invitables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrupoInvitableResponse'
        '403':
          description: Solo el líder puede invitar integrantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Hay un viaje en curso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/grupos/{grupoId}/invitar-desde-grupos:
    post:
      summary: Invitar integrantes desde otros grupos
      description: |
        Crea invitaciones pendientes (`grupo_invitacion`) para todos los miembros de los grupos origen seleccionados (SCRUM-13).
        Excluye al invitador y a quienes ya son miembros del grupo destino.
        Alta masiva en una sola transacción; idempotente ante re-invitaciones.
      operationId: invitarDesdeGrupos
      tags:
        - Grupos
      security:
        - BearerAuth: []
      parameters:
        - name: grupoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitarDesdeGruposRequest'
      responses:
        '201':
          description: Invitaciones registradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitarDesdeGruposResponse'
        '400':
          description: Validación fallida o grupo origen igual al destino
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Solo el líder puede invitar o sin acceso al grupo origen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Grupo destino u origen no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Hay un viaje en curso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token

  parameters:
    XUserId:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      deprecated: true
      description: Reemplazado por BearerAuth

  schemas:
    SyncUsuarioRequest:
      type: object
      required:
        - email
        - nombre
      properties:
        email:
          type: string
          format: email
        nombre:
          type: string
          minLength: 1

    SyncUsuarioResponse:
      type: object
      required:
        - id
        - email
        - nombre
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nombre:
          type: string

    CreateGrupoRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 80
          description: Nombre del grupo (obligatorio)

    GrupoResponse:
      type: object
      required:
        - id
        - nombre
        - fecha_creacion
        - lider_id
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        fecha_creacion:
          type: string
          format: date-time
        lider_id:
          type: string
          format: uuid

    GrupoDetalleResponse:
      allOf:
        - $ref: '#/components/schemas/GrupoResponse'
        - type: object
          required:
            - mi_rol
          properties:
            mi_rol:
              type: string
              enum:
                - lider
                - participante

    GrupoMiembroResponse:
      type: object
      required:
        - id
        - nombre
        - email
        - rol
        - fecha_union
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum:
            - lider
            - participante
        fecha_union:
          type: string
          format: date-time

    CambiarRolMiembroRequest:
      type: object
      required:
        - rol
      properties:
        rol:
          type: string
          enum:
            - lider
            - participante

    CambiarRolMiembroResponse:
      type: object
      required:
        - usuario_id
        - rol
      properties:
        usuario_id:
          type: string
          format: uuid
        rol:
          type: string
          enum:
            - lider
            - participante

    AbandonarGrupoRequest:
      type: object
      properties:
        nuevo_lider_id:
          type: string
          format: uuid
          description: Requerido si el líder abandona y hay otros miembros

    GrupoMutationResponse:
      type: object
      required:
        - accion
      properties:
        accion:
          type: string
          enum:
            - abandonado
            - grupo_eliminado
        grupo_id:
          type: string
          format: uuid

    UsuarioParaInvitarResponse:
      type: object
      required:
        - id
        - nombre
        - email
        - ya_es_miembro
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        email:
          type: string
          format: email
        ya_es_miembro:
          type: boolean
          description: true si ya es miembro o tiene invitación pendiente

    UsuarioInvitableResponse:
      type: object
      required:
        - id
        - nombre
        - email
        - grupos_origen
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        email:
          type: string
          format: email
        grupos_origen:
          type: array
          items:
            $ref: '#/components/schemas/InvitacionGrupoRef'

    InvitarUsuariosRequest:
      type: object
      required:
        - usuario_ids
      properties:
        usuario_ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid

    InvitarUsuariosResponse:
      type: object
      required:
        - invitaciones_creadas
        - invitados
      properties:
        invitaciones_creadas:
          type: integer
          minimum: 0
        invitados:
          type: array
          items:
            type: object
            required:
              - id
              - nombre
            properties:
              id:
                type: string
                format: uuid
              nombre:
                type: string

    GrupoInvitableResponse:
      type: object
      required:
        - id
        - nombre
        - cantidad_integrantes
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        cantidad_integrantes:
          type: integer
          minimum: 0

    InvitarDesdeGruposRequest:
      type: object
      required:
        - grupo_origen_ids
      properties:
        grupo_origen_ids:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            format: uuid

    InvitarDesdeGruposOrigenResult:
      type: object
      required:
        - id
        - nombre
        - invitaciones_creadas
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        invitaciones_creadas:
          type: integer
          minimum: 0

    InvitarDesdeGruposResponse:
      type: object
      required:
        - invitaciones_creadas
        - omitidos_ya_miembros
        - grupos_origen
      properties:
        invitaciones_creadas:
          type: integer
          minimum: 0
        omitidos_ya_miembros:
          type: integer
          minimum: 0
        grupos_origen:
          type: array
          items:
            $ref: '#/components/schemas/InvitarDesdeGruposOrigenResult'

    InvitacionGrupoRef:
      type: object
      required:
        - id
        - nombre
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string

    InvitacionUsuarioRef:
      type: object
      required:
        - id
        - nombre
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string

    InvitacionPendienteResponse:
      type: object
      required:
        - id
        - created_at
        - grupo
        - invitado_por
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        grupo:
          $ref: '#/components/schemas/InvitacionGrupoRef'
        grupo_origen:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/InvitacionGrupoRef'
        invitado_por:
          $ref: '#/components/schemas/InvitacionUsuarioRef'

    ResponderInvitacionRequest:
      type: object
      required:
        - accion
      properties:
        accion:
          type: string
          enum:
            - aceptar
            - rechazar

    ResponderInvitacionResponse:
      type: object
      required:
        - invitacion_id
        - grupo_id
        - grupo_nombre
        - accion
      properties:
        invitacion_id:
          type: string
          format: uuid
        grupo_id:
          type: string
          format: uuid
        grupo_nombre:
          type: string
        accion:
          type: string
          enum:
            - aceptada
            - rechazada
        ya_era_miembro:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string

    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object